+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implement minimum recorder agent for ROS 2 MCAP files, [PR-6](https://github.com/reductstore/reductstore_agent/pull/6)
 - Add code linting with flake8, pep257, and copyright checks, [PR-19](https://github.com/reductstore/reductstore_agent/pull/19)
+- Add support for bucket settings, [PR-22](https://github.com/reductstore/reductstore_agent/pull/22)
 
 [Unreleased]: https://github.com/reductstore/reductstore_agent/compare/v0.1.0...HEAD
  <!-- [0.1.1]: https://github.com/reductstore/reductstore_agent/compare/v0.1.0...v0.1.1 -->